emacs-like window commands
Hooray! We're now about as good as our big brothers when it comes to
windows. We have most of the C-x number commands, except C-x 3 (because
like all ersatz emacsen, we don't have horizontal splits) and the C-x 5
family (because we're in a terminal, so it's a bit silly to have frame
commands)