Optimise the syntax highlighting
Before this patch, the whole buffer was being re-highlighted at every
change. This was fine for normal typing, but pasting in large sections
to a syntax highlighted file (e.g. for pasting code snippets) slowed the
editor to a crawl. Now, the buffer will only be re-highlighted if we can
believe beyond reasonable doubt that it needs it - i.e. if the EOL state
changed. This is similar to how it was done before with the dumb syntax
highlighting from kilo.