Tweak coloff and render code
It still doesn't quite work with UTF-8 shit, but it's a little
better. At this stage, it's not an issue with the windowing code, but
a general issue with coloff. I think it's a confusing overloading of
the coloff variable (which should be an rx-style value).

I hate dealing with rune width and multi-byte characters.