Use visible width for column number on status line (#25)
Update the status line to display the visible width of the characters
before the cursor instead of the number of bytes before the cursor.
These are different numbers for lines with tab characters or multi-byte
UTF-8 characters.  With this change, the Gomacs status line column
number is the same as the GNU Emacs status line column number.