Remove unnecessary use of editorRowRxToCx
editorFindCallback and doReplaceString operate on row.Render, which
is like row.Data except tabs have been converted into spaces.  When
doing that, the indices are an "rx" value (because tabs are counted
as Global.Tabsize characters), which must be converted back to a "cx"
(byte) value.  But I don't see any reason to use row.Render rather
than Row.Data, and one reason not to: searching for spaces will match
tab characters, which is wrong (and the highlight doesn't display
correctly).

doReplaceRegexp is already operating on row.Data and it sets cx to
row.Size, which is already a byte value, so using editorRowRxToCx
seems to serve no purpose.

Aside from the above, the impetus for eliminating editorRowRxToCx in
these locations is that my next commit makes changes to that function
which are not compatible with its usage in nav.go.