Save & restore point in fill-paragraph (#32)
doFillParagraph uses transposeRegion with the FillString callback to
perform the fill operation.  This leaves point (i.e. the cursor) at
the end of the paragraph.  Add routines to save the logical location
of point and to restore point to that location after the fill.  This
allows invoking doFillParagraph without losing your place.

Relies on the fact that the number of "words" (sequences of runes
separated by spaces or newlines) in the paragraph is currently the
the same before & after the fill.

Fixes #29