Allow inserting tabs into prompts (#34)
Most of the time (searching and replacing, for example) we want to
allow the user to insert tabs. Other times, tab will run
autocompletion, and thus we remove tabs from the query. This does have
the side effect that you cannot tab complete on files that have tabs
in their names, but this is a very edge-y edge case we can fix if and
when it breaks someone's workflow.