Make auto-fill-mode less aggressive
Change auto-fill-mode so that it only fills the current line when space
is pressed and point is beyond the fill column.  This is the behavior
used by the GNU Emacs auto-fill-mode.

The previous behavior -- filling the whole paragraph whenever space is
pressed and the line length is beyond the fill column -- is available
with aggressive-fill-mode.

Fixes #30