Force forward progress for query-replace-regexp
A user can specify a regexp (such as 'a*') which matches the zero-length
string.  To avoid an infinite loop for this weird edge case, detect the
condition and force forward progress of one rune per iteration.

With this change, Gomacs and GNU Enmacs do the same thing for this edge
case.

Fixes #36