Fix cursor position with empty lines
When the cursor is on an empty line, the row.RenderSize will be
zero, so row.PrintWCursor won't be called.  The cursor will be
left at its previous position, (0,y).  However, an x-coordinate
of 0 is only correct for the leftmost window and only if there
is no gutter.

Fix this issue by putting the cursor at the start of the line,
accounting for leftward windows and the gutter.